requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
scikit_learn==1.2.1
streamlit==1.18.1
